microservices/user-service/requirements.txt
flask==3.1.0
gunicorn==23.0.0
requests==2.32.3
psycopg2-binary==2.9.9
